게시물 수정, 삭제권한 체크, 로그인 체크, isp include 사용법

+++ src/main/java/com/koreaIT/java/am/ArticleDoDeleteServlet.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
+import java.util.Map;
 
 import com.koreaIT.java.am.config.Config;
 import com.koreaIT.java.am.util.DBUtil;
@@ -15,6 +16,7 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 
 @WebServlet("/article/doDelete")
 public class ArticleDoDeleteServlet extends HttpServlet {
@@ -23,6 +25,8 @@ public class ArticleDoDeleteServlet extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		response.setContentType("text/html; charset=UTF-8;");
+		
+		HttpSession session = request.getSession();
 		
 		Connection conn = null;
 
@@ -33,7 +37,30 @@ public class ArticleDoDeleteServlet extends HttpServlet {
 			
 			int id = Integer.parseInt(request.getParameter("id"));
 			
-			SecSql sql = SecSql.from("DELETE FROM article");
+			SecSql sql = SecSql.from("SELECT * FROM article");
+			sql.append("WHERE id = ?", id);
+			
+			Map<String, Object> articleRow = DBUtil.selectRow(conn, sql);
+			
+			int loginedMemberId = -1;
+			
+			if (articleRow.isEmpty()) {
+				response.getWriter().append(String.format("<script>alert('존재하지 않는 게시물입니다.'); location.replace('list');</script>"));
+				return;
+			}
+			
+			if (session.getAttribute("loginedMemberId") != null) {
+				loginedMemberId = (int) session.getAttribute("loginedMemberId");
+			}
+			
+			if (loginedMemberId != (int) articleRow.get("memberId")) {
+				
+				response.getWriter().append(String.format("<script>alert('권한이 없습니다.'); location.replace('../article/detail?id=%d');</script>", articleRow.get("id")));
+				
+				return;
+			}
+			
+			sql = SecSql.from("DELETE FROM article");
 			sql.append("WHERE id = ?", id);
 			
 			DBUtil.delete(conn, sql);
